Remove duplicates of SqliteWrapper
