+++ QKSMS/src/main/java/com/android/mms/transaction/TransactionService.java
@@ -26,7 +26,6 @@ import android.content.Context;
 import android.content.Intent;
 import android.content.IntentFilter;
 import android.database.Cursor;
-import android.database.sqlite.SqliteWrapper;
 import android.net.ConnectivityManager;
 import android.net.NetworkInfo;
 import android.net.Uri;
@@ -46,8 +45,9 @@ import com.google.android.mms.pdu_alt.NotificationInd;
 import com.google.android.mms.pdu_alt.PduHeaders;
 import com.google.android.mms.pdu_alt.PduParser;
 import com.google.android.mms.pdu_alt.PduPersister;
-import com.moez.QKSMS.mmssms.Utils;
 import com.moez.QKSMS.R;
+import com.moez.QKSMS.common.SqliteWrapper;
+import com.moez.QKSMS.mmssms.Utils;
 
 import java.io.IOException;
 import java.util.ArrayList;
