+++ QKSMS/src/main/java/com/google/android/mms/pdu_alt/PduPersister.java
@@ -33,7 +33,6 @@ import android.telephony.PhoneNumberUtils;
 import android.telephony.TelephonyManager;
 import android.text.TextUtils;
 import android.util.Log;
-
 import com.google.android.mms.ContentType;
 import com.google.android.mms.InvalidHeaderValueException;
 import com.google.android.mms.MmsException;
@@ -41,7 +40,7 @@ import com.google.android.mms.util_alt.DownloadDrmHelper;
 import com.google.android.mms.util_alt.DrmConvertSession;
 import com.google.android.mms.util_alt.PduCache;
 import com.google.android.mms.util_alt.PduCacheEntry;
-import com.google.android.mms.util_alt.SqliteWrapper;
+import com.moez.QKSMS.common.SqliteWrapper;
 import com.moez.QKSMS.mmssms.Utils;
 
 import java.io.ByteArrayOutputStream;
