+++ QKSMS/src/main/java/com/moez/QKSMS/common/utils/MessageUtils.java
@@ -30,7 +30,6 @@ import android.content.DialogInterface.OnClickListener;
 import android.content.Intent;
 import android.content.res.Resources;
 import android.database.Cursor;
-import android.database.sqlite.SqliteWrapper;
 import android.drm.DrmStore;
 import android.media.CamcorderProfile;
 import android.media.RingtoneManager;
@@ -61,10 +60,12 @@ import com.google.android.mms.pdu_alt.PduPart;
 import com.google.android.mms.pdu_alt.PduPersister;
 import com.google.android.mms.pdu_alt.RetrieveConf;
 import com.google.android.mms.pdu_alt.SendReq;
-import com.moez.QKSMS.common.LogTag;
-import com.moez.QKSMS.common.MmsConfig;
 import com.moez.QKSMS.QKSMSApp;
 import com.moez.QKSMS.R;
+import com.moez.QKSMS.common.LogTag;
+import com.moez.QKSMS.common.MmsConfig;
+import com.moez.QKSMS.common.SmsHelper;
+import com.moez.QKSMS.common.SqliteWrapper;
 import com.moez.QKSMS.common.TempFileProvider;
 import com.moez.QKSMS.common.google.ThumbnailManager;
 import com.moez.QKSMS.common.google.UriImage;
@@ -72,7 +73,6 @@ import com.moez.QKSMS.data.Contact;
 import com.moez.QKSMS.model.MediaModel;
 import com.moez.QKSMS.model.SlideModel;
 import com.moez.QKSMS.model.SlideshowModel;
-import com.moez.QKSMS.common.SmsHelper;
 import com.moez.QKSMS.ui.dialog.AsyncDialog;
 import com.moez.QKSMS.ui.messagelist.MessageColumns;
 import com.moez.QKSMS.ui.messagelist.MessageItem;
