+++ QKSMS/src/main/java/com/moez/QKSMS/data/Conversation.java
@@ -7,7 +7,6 @@ import android.content.ContentUris;
 import android.content.ContentValues;
 import android.content.Context;
 import android.database.Cursor;
-import android.database.sqlite.SqliteWrapper;
 import android.net.Uri;
 import android.os.AsyncTask;
 import android.provider.BaseColumns;
@@ -19,20 +18,20 @@ import android.provider.Telephony.Threads;
 import android.provider.Telephony.ThreadsColumns;
 import android.text.TextUtils;
 import android.util.Log;
-
 import com.android.mms.transaction.MmsMessageSender;
 import com.google.android.mms.pdu_alt.PduHeaders;
-import com.moez.QKSMS.QKSMSAppBase;
-import com.moez.QKSMS.mmssms.Utils;
-import com.moez.QKSMS.common.LogTag;
 import com.moez.QKSMS.QKSMSApp;
+import com.moez.QKSMS.QKSMSAppBase;
 import com.moez.QKSMS.R;
-import com.moez.QKSMS.common.google.DraftCache;
-import com.moez.QKSMS.receiver.UnreadBadgeService;
-import com.moez.QKSMS.common.utils.AddressUtils;
-import com.moez.QKSMS.common.utils.PhoneNumberUtils;
+import com.moez.QKSMS.common.LogTag;
 import com.moez.QKSMS.common.NotificationManager;
 import com.moez.QKSMS.common.SmsHelper;
+import com.moez.QKSMS.common.SqliteWrapper;
+import com.moez.QKSMS.common.google.DraftCache;
+import com.moez.QKSMS.common.utils.AddressUtils;
+import com.moez.QKSMS.common.utils.PhoneNumberUtils;
+import com.moez.QKSMS.mmssms.Utils;
+import com.moez.QKSMS.receiver.UnreadBadgeService;
 
 import java.util.ArrayList;
 import java.util.Collection;
