+++ QKSMS/src/main/java/com/moez/QKSMS/model/VideoModel.java
@@ -20,26 +20,24 @@ package com.moez.QKSMS.model;
 import android.content.ContentResolver;
 import android.content.Context;
 import android.database.Cursor;
-import android.database.sqlite.SqliteWrapper;
 import android.net.Uri;
 import android.provider.MediaStore.Images;
 import android.provider.Telephony.Mms.Part;
 import android.text.TextUtils;
 import android.util.Log;
 import android.webkit.MimeTypeMap;
-
 import com.android.mms.dom.events.EventImpl;
 import com.android.mms.dom.smil.SmilMediaElementImpl;
 import com.google.android.mms.ContentType;
 import com.google.android.mms.MmsException;
 import com.google.android.mms.smil.SmilHelper;
-import com.moez.QKSMS.exception.ContentRestrictionException;
-import com.moez.QKSMS.common.LogTag;
 import com.moez.QKSMS.QKSMSApp;
+import com.moez.QKSMS.common.LogTag;
+import com.moez.QKSMS.common.SqliteWrapper;
 import com.moez.QKSMS.common.google.ItemLoadedCallback;
 import com.moez.QKSMS.common.google.ItemLoadedFuture;
 import com.moez.QKSMS.common.google.ThumbnailManager;
-
+import com.moez.QKSMS.exception.ContentRestrictionException;
 import org.w3c.dom.events.Event;
 import org.w3c.dom.smil.ElementTime;
